requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pyjwt==2.9.0
aiofiles==24.1.0
